backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.*
python-dotenv==1.0.*
redis==5.*
SQLAlchemy==2.0.*
psycopg[binary]==3.1.*
httpx==0.27.*
apscheduler==3.10.*
paramiko
